Eleven bytes alpha
This is alpha of very popular game 2048. It is vary stupid and not very logical. It is have a lot of copypasted code, but it is wokring. Right now there is two classes: GUI and Map. Second one resposible for game logic and storing information about cells and first one for showing it.