Logic of movement and merge change.
Change logic of this two methods. Know it pocessing in one method not four. Now it is looking more beutifull and logical. Also I write all documentation to some methods and refactor Map.java. And one more small change: it is score system. Now it is just adding value of new elements to the score.