GitHubBot Working Beta
I hope it is gonna work. Wonder why and what. Well if you this message in my twitter everyting is going fine and my bot posted it for me, probably with love. In this version I update some small tips and change time of updates comming to github. Well... there is only one way to find out. Oh... wait. It is doesn't work. My bot making to much updates.